[NEW] 2019/25 (solved 2019-12-25)
This is not a general solution by any means.  It's specific to my input,
in that it relies on knowing, in advance, the names of objects, which of
them are safe to take, where each is located, how to get there, and also
the subset of objects needed to pass the security checkpoint -- although
it actually retrieves all the objects anyway, including the unused ones.