Prefer `pop` operator to `Iter.popcount (C ver P)`
It's redundant and incoherent to run a popcount over a sequence of bytes
that have already been verified to have a non-zero value.  Additionally,
constructing the result of the `ver` operation is wasted work (as is the
job of subsequently traversing it) when all we need is the member count.